pyproject.toml
[project]
name = "hairbrush"
version = "0.1.0"
description = "An example Django SaaS app."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "django-environ>=0.13.0",
    "django-health-check>=4.4.0",
    "gunicorn>=25.3.0",
    "psycopg[binary,pool]>=3.3.3",
    "uvicorn[standard]>=0.44.0",
    "whitenoise[brotli]>=6.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.ruff]
target-version = "py314"
line-length = 120
